refactor(networking): move tests to new file
